chore(test): update test mock
