+++ src/components/presentations/ArticleCard/__test__/ArticleCard.test.jsx
@@ -1,19 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import ArticleCard from '../Article';
+import mock from '../../../../utils/testMocks';
 
-const article = {
-  id: 1,
-  category: 'Health',
-  title: 'World’s Smallest Biggest Killer',
-  author: 'Andrew Moffins',
-  date: 'March 26 2019',
-  read: '1 min read',
-  paragraph:
-    'The world’s biggest killers are in fact so tiny - the mosquito. Worldwide there are 3500 known species of mosquito. Which are the deadliest the mosquito. Worldwide there are 3500 known species of mosquito...',
-  image:
-    'https://res.cloudinary.com/walsam/image/upload/v1556790931/robotics.jpg',
-};
+const { article } = mock;
 
 describe('<ArticleCard />', () => {
   it('renders article', () => {
